+++ packages/runtime-tools/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @module-federation/runtime-tools
 
+## 0.11.4
+
+### Patch Changes
+
+- @module-federation/runtime@0.11.4
+- @module-federation/webpack-bundler-runtime@0.11.4
+
 ## 0.11.3
 
 ### Patch Changes